port the helper scripts to python 3
check_generator.py additionally takes the path to vamdctap's generators.py
from VAMDCTAP_GENERATORS.

+++ strip_db.py
@@ -6,7 +6,7 @@ from sqlite3 import dbapi2 as sqlite
 
 FULL = 'dict.sqlite'
 BARE = 'dict_bare.sqlite'
-os.unlink(BARE)
+if os.path.exists(BARE): os.unlink(BARE)
 TABS1 = [ 'browse_keyword', 'browse_usage', 'browse_keyword_usage' ]
 TABS2 = [ tab.replace('browse_','') for tab in TABS1 ]
 
@@ -14,7 +14,7 @@ conn = sqlite.connect(BARE)
 curs = conn.cursor()
 curs.execute("ATTACH '%s' AS orig;"%FULL)
 
-for TAB1,TAB2 in map(None,TABS1,TABS2):
+for TAB1,TAB2 in zip(TABS1,TABS2):
     sql = "SELECT sql FROM orig.sqlite_master WHERE name='%s';"%TAB1
     create, = curs.execute(sql).fetchone()
     curs.execute(create.replace('browse_',''))
